Add test for hasBuildings method in GameBoard class
This commit adds a test method hasBuildings_ShouldReturnTrue_WhenPropertiesInGroupHaveBuildings to the test class. This test verifies that the hasBuildings method returns true when properties in the specified group have buildings.

The test manually adds buildings to a RED property in the setup method. It then arranges by setting up a group color and acts by calling hasBuildings with this color group. Finally, it asserts that the result is true, indicating buildings are present in the group.