Add canCollectRent method to MortgageManager class
This commit adds the canCollectRent method to the MortgageManager class. This method checks if rent can be collected on properties of a specific color group.

The method checks if any properties in the specified color group are mortgaged. If no properties in the group are mortgaged, it returns true indicating rent can be collected. Otherwise, it returns false.

Add enableBuildingOnGroup method to MortgageManager class

This commit adds the enableBuildingOnGroup method to the MortgageManager class. This method enables building on a property group if no properties in that group are mortgaged.

The method checks if any properties in the specified color group are mortgaged. If no properties in the group are mortgaged, it enables building on the group.